+++ limix/qtl/__init__.py
@@ -12,10 +12,10 @@ Classes
 """
 
 
-from ._st_scan import st_scan
-from ._st_iscan import st_iscan
-from ._st_sscan import st_sscan
 from ._model import QTLModel
-from ._gwas import GWAS_MTLMM
+from ._mt_scan import mt_scan
+from ._st_iscan import st_iscan
+from ._st_scan import st_scan
+from ._st_sscan import st_sscan
 
-__all__ = ["st_scan", "st_iscan", "st_sscan", "QTLModel", "GWAS_MTLMM"]
+__all__ = ["st_scan", "st_iscan", "st_sscan", "QTLModel", "mt_scan"]
